% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_Size_for_EV.R
\name{ss_citl}
\alias{ss_citl}
\title{Minimum sample size for external validation of a clinical prediction model with a continuous outcome}
\usage{
ss_citl(varY, R2, width, alpha)
}
\arguments{
\item{varY}{Variance of the observed Yi}

\item{R2}{Expected coefficient of determination (R2) of CITL}

\item{width}{Expected Width of confidence interval of CITL}

\item{alpha}{Level of Significance}
}
\value{
A html file contains minimum sample size
}
\description{
Precise estimate of CITL (calibration-in-the-large)
}
\examples{

#Eg. To target SE of CITL model of 2.55 (width = 10)at 95\%
#confidence interval , Assuming R2 CITL = R2val = 0.5 and
# variance of the observed Yi is 400
ss_citl( R2= 0.5,width = 10,alpha = 0.05, varY = 400)

#for different combinations of parameters
ss_citl( R2= c(0.4,0.7,0.2),width =c( 1.17),alpha = 0.05, varY = 400)
}
\references{
Archer L, Snell KIE, Ensor J, Hudda MT, Collins GS, Riley RD. Minimum sample size for external validation of a clinical prediction model with a continuous outcome. Stat Med. 2020;40:133-46.
}
\seealso{
ss_R2val, ss_cal_slope , ss_res_var , ss_sens , ss_spec
}
