% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_Size_for_EV.R
\name{ss_R2val}
\alias{ss_R2val}
\title{Minimum sample size for external validation of a clinical prediction model with a continuous outcome}
\usage{
ss_R2val(R2val, width, alpha)
}
\arguments{
\item{R2val}{Assumed coefficient of determination}

\item{width}{Expected width of confidence interval}

\item{alpha}{Level of Significance}
}
\value{
A html file contains minimum sample size
}
\description{
Precise estimate of R2val
}
\examples{
#Eg.To target a 95\% confidence interval for R2val that has a
#narrow width of about 0.1 i.e SE of R2val will be 0.0255 and
#assuming R2val is 0.5.
ss_R2val( R2val= 0.5,width=0.1,alpha=0.05 )

#for different combinations of parameters
ss_R2val( R2val= c(0.6,0.9),width=c( 0.1),alpha=c( 0.01,0.05) )
}
\references{
Archer L, Snell KIE, Ensor J, Hudda MT, Collins GS, Riley RD. Minimum sample size for external validation of a clinical prediction model with a continuous outcome. Stat Med. 2020;40:133-46.
}
\seealso{
ss_citl, ss_cal_slope , ss_res_var , ss_sens , ss_spec
}
