% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_Size_for_EV.R
\name{ss_res_var}
\alias{ss_res_var}
\title{Minimum sample size for external validation of a clinical prediction model with a continuous outcome}
\usage{
ss_res_var(alpha, max_MOE)
}
\arguments{
\item{alpha}{Level of Significance}

\item{max_MOE}{Multiplicative margin of error}
}
\value{
A html file contains minimum sample size
}
\description{
Precise estimates of residual variances (small multiplicative margin of error (MMOE) around the true value)
}
\examples{
#Eg.Sample size for margin of error of within 10\% (1.0 <=MMOE <=1.1)
# of the true value at 95\% confidence level.
ss_res_var( max_MOE=1.1, alpha=0.05)

#for different combinations of parameters
ss_res_var( max_MOE=c(1.1,1.2,1.3), alpha=c(0.05,0.01))
}
\references{
Archer L, Snell KIE, Ensor J, Hudda MT, Collins GS, Riley RD. Minimum sample size for external validation of a clinical prediction model with a continuous outcome. Stat Med. 2020;40:133-46.
}
\seealso{
ss_citl, ss_cal_slope , ss_R2val , ss_sens , ss_spec
}
